Extract public key hash from `PubScript` received by the blockflow node
Lots of changing in that commit to achieve the title. As our model and
the one from blockflow is diverging, we need to create two models, might
seems overkill now as it's mostly the same info, but this allow in the
future to easily change the explorer's model, while the one from
blockflow don't change, it's  good practice.