Add `txHashRef`, `address` and `amount` to Input
This implies lot of change, basically we can't work anymore only with
the api models, as we can't go directly from the protocol models to the
ap ones. So we go from protocol models to entity ones and the api models
are used only to respond to the endpoints, which is a better practice.

Theres still a bit of cleaning to do to clearly separate all the models,
but it can be done latter.