Add `mainChain` field in `TransactionEntity`
We had a bug in the case where a transaction was in two blocks, one on the
main chain and the other on a fork.

We thus had two times the transaction in db, so when we were listing the
transactions for then concerned address, we were getting two times the
tx.

We are now filtering by main chain the same way we do for inputs and
outputs.