Improve TokenCirculationService
The workflow is now as follow:

- Find the minimum block day between latest block of each chain
  We know then that the previous day has been completed and no more
  block will be added to that day.
- Find the latest updated day.
- Build the range of days between the latest updated day and the latest
  completed day.
- If the range is not empty: update the token circulation for each of those days.

With this version we make sure that we only update the circulation
number for a day where we have all the blocks synced for that day.
That was not the case with the previous version, where we were computing
the circulation while the blocks were not synced.

Currently we sync every minutes, but that's not a problem as we always
check if there are days to sync, if not, we just do nothing and check
again in 1 minute.

An obvious improvement would be to wait for a all day, but currently we
can't do it as the `syncPeriod` of `SyncService` has to be a fixed
value.