Introduce `SanityChecker`
Currently a bit "brute-force", we take the highest block on each chain
and go down through the parents up until the genesis blocks, checking
every time that the block is the main_chain one.

Also note that unless we have a missing block, which then require to
call the node, all the sanity check happen in a `DBAction`

We can't stop yet the sanity check, latter we'll have an action to
start/stop it.