Cache the latest block for each chain
We have here the first case where we clearly need to differentiate the
behavior between read/write and read-only app.

In the first case we can keep in memory the latest blocks, but we have
to store those data in db, so the read-only app can also access it.

We could think if latter we want to introduce some other communication
channel between the two types of app for such data, maybe with a simple
websocket. But for now I think postgres is fine for this kind of
feature.