Test `getOutputsQuery`
In previous commit we fixed the following bug:

When we were getting an output, we were searching if that output was
spent, so we were joining on the inputs table, but not the main-inputs
table. So when there were two inputs for that output (one in main-chain
and the other not), the resulting join was giving twice the same output.