Show timestamp convertion issue
Using the default behavior of java.sql.timestamp works only when using
slick only, as soon as we have raw sql query, it doesn't work.

Because with our current `TimeStamp` custom type, the store timestamp is
actually stored in db as a local date, so shifted by +1 hour for example
here in Switzerland.

Next commit will fix it by making sure the data written in db are the
correct UTC timestamp