Rename `ChainHeight` to `PerChainValue`
The goal is to have a type that is reusable for other "per chain" info,
in a coming PR we will return the nb of txs per chain and we will reuse
the same type.

The same process will be applied for the charts, but with another type.

The overall purpose is to simplify integration with FE.