Refactor Scheduler to always a `Future` while the loop is running.
Prior to this commit we had two ways:
 - `scheduleLoop` was completing the initial call as a future and then
   fire-and-forgetting the loop
 - `scheduleLoopAndForget` was returning `Unit`

Both approaches have the disadvantage to "loose" our `Future` and thus
we cannot interact with them anymore.
With this commit every schedule function is returning a `Future` and run
as long as the loop is scheduled, in case of error in the loop, we don't
recover as before, we fail the future and let the scheduler user decide
what to do.