Introduce `InputUpdateQueries`
The previous strategy to update data that require to find the
`output_ref` of input was not good. When we were inserting a block, we
were trying to join inputs and outputs to make the updates, if we were
not finding the corresponding output, we were delaying the update of the
inputs (`inputsToUpdate` values) and hoping it would succeed the second
time when blocks for the given time range were synced.

This was a bit ugly and still a bit risky as we could still miss an
output on second try.

The new strategy found uses the new `output_ref_tx_hash` field, which is NULL when we
insert the inputs.
Once we have sync all blocks of the time range, we call the
`InputUpdateQueries.updateInputs` function, which try to update every input where `output_ref_tx_hash IS NULL`.

The advantage is that no inputs could be left behind, even if we miss an
output, eventually this output will be synced and on a latter run of
`updateInputs` the input will be updated.

Once the corresponding output is found, we directly update what's rely
on it, like the `transaction_per_addresses` table. `transactionally` is used
to make sure we update everything or nothing.