Optimize output query
We were actually joining the inputs to get the `spent` field.
This field is not that important for the front-end, it just allows to
create a link to the transaction where the output is spent.
We already get that value in the `spent_finalized` field, it has a bit
of delay since we need the output to be finalized, but in a sense it's
more accurate to wait for it and this improve a lot the query as we
don't need to join.