Cache latest blocks based on pagination
The goal is to improve the home page of our explorer.

We trust here the eviction algorithm of Caffeine:

[tinyLFU](https://dl.acm.org/doi/pdf/10.1145/3149371)

As the first page of the explorer is the one that will be called almost
all the time compared to any other pages, it should almost never be
evicted by the cache.

Resolves: #151