Migrate to `vertx`
I tried to have a readable diff, I redo what I did in the node, keeping
same wording as akka-http-testkit help.

Unfortunatelly with `vertx` we have to start the server: `server.listen`
in order to test the routes, we can't do as in akka.

So know for all the `web.*ServerSpec` I start once the server on top of
the file and all test are using the same instance.