Introduce `CreateSubContractEvent` types
Naming comes from the full node and the hardcoded [createContractEventIndex](https://github.com/alephium/alephium/blob/master/protocol/src/main/scala/org/alephium/protocol/vm/package.scala#L47)

[A contract write a log on creation](https://github.com/alephium/alephium/blob/c84dc660b8e7648827d16c3caef78f87205405fc/protocol/src/main/scala/org/alephium/protocol/vm/Frame.scala#L373-L408)
from the `contractCreationEventFields` function, we know that the firt
address in the log fields is the sub-contract address, while the second
address it the parent one.
We can ignore the first field `createContractEventIndex`, as it's discarded before the log.

In the explorer-backend, when we receive and event, we can check if it's
a contract creation event with the `createContractEventIndex` contract
address and then check if there are 2 fields, which means it's a
sub-contract creation.