Add `/addresses/<address>/tokens-balance` endpoint
On the explorer front-end, we start to have `too many request` on
addresses pages with lots of tokens, as currently you need to loop and
call the balance endpoint for each token.

This endpoint allows to do only one call to get all tokens info needed.

This is currently quite a naive approach, also we need to handle the
case where balance is 0, we shouldn't return the tokens in those case.
But this requires some work to do things in an efficient way.