Use `Flowable.fromCompletionStage` instead of `fromFuture`
This one seems to be a game changer.

`fromFuture` is actually performing some blocking operations and is not
recommended.
This little change improve a lot the export of txs. With the former we
were reaching 600k/s, up to 700k/s when jvm and db was hot.
With the second one, it reaches around 1400k/s and up to 2700k/s when
hot on my various tests.