Clean all our functions name
During our long phase of optimizing our queries, we were keeping old
function and introducing new one in parallel, to test they were
returning the same result.

We were always adding some `SQL` and/or `NoJoin` to be explicit.

Now that it's stable since a long time, we already removed the old
function, but didn't adapt the naming.

This PR clean all our names, removing all the `SQL/NoJoin` that made the
code dirty and confusing.