Add `spent_timestamp` to outputs
This field will latter be used to optimize our `TokenSupplyService`.

The migrations of the data is done by reseting the
`last_finalized_input_time`, the `FinalizerService` will then reprocess
all inputs/outputs and update the `spent_timestamp` field.
The `spent_finalized` field, used to optimized the balance query aren't
deleted, it will just be re-updated with the same value.

This process take a bit of time, 3-4 hours on my machine, but it's done
in the background and the `spent_timestamp` isn't used for now so it's
safe.