Update token supply only once a day
Currently when the update was done, we were restarting after 1 minutes.
That was quite stupid I must admit, knowing from our previous commits
how slow our queries were. Especially because the token supply value is
only available once a day.

With this commit we now compute the token supply only once a day, I
propose at 2AM UTC, so we are sure the finalization time has passed and
we can safely compute the supply of the previous day.