Add tokens and contracts data to `ExplorerSpec`
We aren't testing thecorrectness of the data as it's complicated to
generate coherent data.
But adding such test in `ExplorerSpec` allows to test the all pipeline
from the syncing to the api, passing by the DB etc.

I had lot of issue with writing those tests, due primarily to our tokenId
and contractId that were not generated correct data, their group were
always wrong.