Remvove deps table and store serialized deps in block_headers table
The deps table was actually the biggest table, around 29Go at the time
of this commit, while we were not using it, only to add it to
block_entry.

When first building the explorer-backend, the idea was to follow the
sql's good practice and store the deps array in a relation table, so we
could make queries and join agains the deps table.

In the end, deps are only there for display, so better to store them as
a serialized data directly in the block table.