Return full `BlockEntry` when requesting by hash
We were only returning the Lite one for now, as the FE didn't wanted to
get the transactions in the block, as it's complicated to have
pagination.

We are still not returning the transactions, but we give all other info
of the blocks, so FE can slowly show more info.

Transactions are still requested through another endpoint.