Correct coinbase value in `transaction_per_address` table
Previously, when updating inputs, the coinbase value from the output_ref
was incorrectly used in the transaction_per_address table.
This was incorrect because the output_ref's coinbase value is not the
actual transaction of the input. Additionally, a transaction with inputs
cannot be a coinbase transaction.

A migration is required to correct the coinbase values in the database.

Before this fix, API users may experience duplicate transactions for certain addresses.