Use `BlockCache.latestBlocks` to speed up syncing
The block cache was introduced later and we never use it for the syncing
process.

So when starting the app and before every 5second sync, we were using an
heavy query to find the latest synced block.

This explain why it took a lot of resources on boot up and also that the
syncing could be a bit slow.