Reduce call to node when synced
Every 5 seconds we were calling the node to know the latest block
timestamp to sync.

Once we are up-to-date we can simply use `now` to avoid unnecessary
call.

If we are not up-to-date it's safer to ask the node. If for example you
are syncing explorer-backend while the node is also re-syncing from
scratch, using `now` will create a lot of call returning empty blocks.