Start metric and transaction cache without loading data
We were waiting to populate the cache on start, which means that the all
app, like http server and sync services were waiting on those cache.

We now start directly, but still expire the cache so the reload will be
trigger on next call.

The only drawback is that the very first call on those cache will give a
0 value, but I think we can live with that. This avoid unnecessary
queries if those cache are never requested.