Prevent `SttpBackend` to throw error on close
When stopping the `MarketService`, we have to close the `SttpBackend`
unfortunately we don't have much control over it and when it's closed
while doing a request, it's throwing an error and we were logging that
error.

It's correct in a way, but it was confusing on the reason why the
explorer-backend was stopping.

For example if the node stop while the market service is updating its
caches, an error was logged, saying the caches couldn't get updated.

This was confusing and let user think the market service was the reason
for the explorer-backend to stop.

We are now making sure that if we catch an error while we the service is
stopped or not running, we don't throw the error.

On the other hand if the service is running, we do want to log the
error.

Resolves: #603