Introduce a Default group config
I know it's no ideal, but with the recent change in full node `ApiModel`
an implicit group config is required everywhere, due to the new
`Address.fromBase58` that needs that group config.

In full node, only the endpoints logic needs to define this group
config, while on explorer-backend side, we use the `fromBase58` almost
everywhere, as it's required to read an address from database.

So we would have to change almost all code base to be able to nicely
pass the group config, while it's not a dynamic value. It's always the
same once booted.

Maybe later I will come back to this.