Change address tapir schema to generate optinal group
In case the address is a P2PK, we extract the optional group so it can
be passed further.

As with the current `Address` reader, once the address is built with the
passed group, we can't differentiate if the address was created with the
default group @x, or using the groupless format.