Update node to v3.15.1
This commit needs to adapt the new way of handling half-decoded
addresses.

Now all the half-decoded addresses are managed in `org.alephium.api.model.Address`
and are used only at `api` level, while the complete addresess are still
defined in `protocol` and are used in the rest of the code base.

We used the same strategy:
  * `ApiAddress` are used as api entry, to let user choose for anykind
    of address
  * `Address` from protocol is used as DB model, as it's the type
    received from the node when syncing
  * `GrouplessAddress` an new wrapper used in DB for all
    `groupless_address` column, which help us to avoid inserting
    non-groupless address in those columns.