Avoid to much price fetching of invalid tokens
At the time of the commit, we have around 220 tokens in our token-list.

Only around 25 tokens have a price and enough liquidity.

The goal of this commit is to avoid to requesting price for tokens we
know aren't valid.

We still re-try every time we refresh the token-list (currently every
12h) to see if the price is now valid.