Optimize `BlockFlowSyncService.getRemoteMaxTimestamp`
We were previously considering that EB was synced if latest synced block
was 10s ago, otherwise we were contacting the node to know the latest
block's timestamp.

We are now using the `blockFlowFetchMaxAge` value, set to 30min in the
node.

If our latest block is younger than (now - blockFlowFetchMaxAge), we
don't need to call the node as we anyway can sync all remaining blocks
in one call.