Ensure transaction inputs return complete output data
There is a rare edge case when querying transactions that have just been inserted,
but whose inputs have not yet been updated by `InputUpdateQueries`.
In this scenario, the API returns the transactions, but the inputs are missing output data such as address and amount.
This makes it difficult for applications to process these transactions.

To address this, the API should only return transactions with complete data.
For this edge case (which typically lasts around 300ms between insert and update),
when querying transactions, we check if input data is complete—which is almost always the case.
If not, we retrieve the corresponding output so users always receive complete transaction data.

If for some reason we can't find the output, we'll return then
incomplete input.