Fix incorrect field order in SQL queries due to use of SELECT *
We were using `SELECT *` in several queries and mapping the results directly to entity models. This approach is fragile and led to issues on testnet/mainnet.

For example, when we added the `conflicted` field to the transaction entity (positioned below `main_chain` in the model), a fresh database created the column in the expected order. However, in testnet/mainnet, where `conflicted` was added via migration, the column was placed at the end. As a result, `SELECT *` returned columns in the wrong order, causing `CustomGetResult` to misalign fields.

This commit replaces `SELECT *` with explicit column selection to ensure correct field mapping regardless of column order in the database schema.

This was also not catch by our test as some models where not creating all data (see previous commit).